refactor(mcp): split MCP and ComputerUse into separate modules
- Move ComputerUse class from mcp.py to new tools.py file
- Update imports in __init__.py and lmcp.py to reflect new module structure
- Remove deprecated import in mcp.py
