+++ lybic/lmcp.py
@@ -14,7 +14,8 @@ from lybic import MCP,ComputerUse
 # pylint: disable=unused-import
 
 import warnings
-from .mcp import MCP, ComputerUse
+from .mcp import MCP
+from .tools import ComputerUse
 
 warnings.warn(
     "The 'lybic.lmcp' module has been moved to 'lybic.mcp' and is deprecated. It will be removed in version 1.0. Please update your imports.",
